refactor(tests): extract shared helpers to package-level helper_test.go files
- [+] Document helper_test.go usage in AGENTS.md
- [+] Move CLI helpers (collectDomains, newCheckCmd, etc.) to internal/cli/helper_test.go
- [+] Move Nawala DNS helpers (startTestDNSServer, etc.) to src/nawala/helper_test.go
- [+] Remove duplicated helpers from cli/check_test.go and nawala/*_test.go files
- [+] Update test calls (e.g., startSimpleDNSServer → startNormalDNSServer)
