+++ internal/cli/check_test.go
@@ -8,34 +8,12 @@ package cli
 import (
 	"context"
 	"fmt"
-	"net"
 	"os"
 	"path/filepath"
 	"strings"
 	"testing"
 	"time"
-
-	"github.com/miekg/dns"
-	"github.com/spf13/cobra"
 )
-
-// collectDomains is a helper wrapper for tests to consume streamDomains into a slice.
-func collectDomains(args []string, filePath string) ([]string, error) {
-	ctx := context.Background()
-	in := make(chan string)
-
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- streamDomains(ctx, args, filePath, in)
-	}()
-
-	var domains []string
-	for d := range in {
-		domains = append(domains, d)
-	}
-
-	return domains, <-errCh
-}
 
 func TestCollectDomains_ArgsOnly(t *testing.T) {
 	domains, err := collectDomains([]string{"a.com", "b.com"}, "")
@@ -258,19 +236,6 @@ func TestCollectDomains_IDNAConversion(t *testing.T) {
 	})
 }
 
-// newCheckCmd creates a fresh check command for isolated testing.
-func newCheckCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:  "check [domains...]",
-		Args: cobra.ArbitraryArgs,
-		RunE: runCheck,
-	}
-	cmd.Flags().StringP("file", "f", "", "path to a .txt file with one domain per line")
-	cmd.Flags().StringP("output", "o", "", "write results to a file instead of stdout")
-	cmd.Flags().StringSlice("format", []string{"text"}, "output format (text, json, html, xlsx)")
-	return cmd
-}
-
 func TestRunCheck_NoDomains(t *testing.T) {
 	saved := configPath
 	configPath = ""
@@ -375,34 +340,6 @@ func TestRunCheck_NoServers(t *testing.T) {
 	}
 }
 
-func createMockDNSServer(t *testing.T) (string, func()) {
-	conn, err := net.ListenUDP("udp", &net.UDPAddr{IP: net.ParseIP("127.0.0.1"), Port: 0})
-	if err != nil {
-		t.Fatal(err)
-	}
-	done := make(chan struct{})
-	go func() {
-		buf := make([]byte, 512)
-		for {
-			n, addr, err := conn.ReadFromUDP(buf)
-			if err != nil {
-				close(done)
-				return
-			}
-			if n >= 12 {
-				// Create a valid DNS response (Standard query response, No error)
-				resp := append([]byte(nil), buf[:n]...) // copy request
-				resp[2] |= 0x80                         // Set QR bit (Response)
-				_, _ = conn.WriteToUDP(resp, addr)
-			}
-		}
-	}()
-	return conn.LocalAddr().String(), func() {
-		_ = conn.Close()
-		<-done
-	}
-}
-
 func TestRunCheck_Success(t *testing.T) {
 	mockAddr, cleanup := createMockDNSServer(t)
 	defer cleanup()
@@ -454,24 +391,6 @@ func TestRunCheck_MultipleFormatFlags(t *testing.T) {
 	}
 }
 
-// collectDomainsCtx is like collectDomains but accepts a context,
-// enabling tests for context-cancellation paths.
-func collectDomainsCtx(ctx context.Context, args []string, filePath string) ([]string, error) {
-	in := make(chan string)
-
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- streamDomains(ctx, args, filePath, in)
-	}()
-
-	var domains []string
-	for d := range in {
-		domains = append(domains, d)
-	}
-
-	return domains, <-errCh
-}
-
 // TestCollectDomains_ContextCancelledDuringArgs covers the ctx.Err() check
 // inside the positional-args loop of streamDomains (check.go line 140-142).
 func TestCollectDomains_ContextCancelledDuringArgs(t *testing.T) {
@@ -506,56 +425,6 @@ func TestCollectDomains_ContextCancelledDuringFileScan(t *testing.T) {
 	if err != context.Canceled {
 		t.Errorf("expected context.Canceled, got: %v", err)
 	}
-}
-
-// createSlowDNSServer starts a proper miekg/dns server that delays before
-// responding, used to guarantee context-deadline expiration during a DNS query.
-// Uses dns.Server for cross-platform reliability (matching the SDK internal tests).
-func createSlowDNSServer(t *testing.T, delay time.Duration) (string, func()) {
-	t.Helper()
-
-	handler := dns.HandlerFunc(func(w dns.ResponseWriter, r *dns.Msg) {
-		time.Sleep(delay)
-		m := new(dns.Msg)
-		m.SetReply(r)
-		m.Answer = append(m.Answer, &dns.A{
-			Hdr: dns.RR_Header{
-				Name:   r.Question[0].Name,
-				Rrtype: dns.TypeA,
-				Class:  dns.ClassINET,
-				Ttl:    60,
-			},
-			A: net.ParseIP("93.184.216.34"),
-		})
-		_ = w.WriteMsg(m)
-	})
-
-	pc, err := net.ListenPacket("udp", "127.0.0.1:0")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	server := &dns.Server{
-		PacketConn: pc,
-		Handler:    handler,
-	}
-
-	started := make(chan error, 1)
-	go func() {
-		server.NotifyStartedFunc = func() { started <- nil }
-		if err := server.ActivateAndServe(); err != nil {
-			select {
-			case started <- err:
-			default:
-			}
-		}
-	}()
-
-	if err := <-started; err != nil {
-		t.Fatal(err)
-	}
-
-	return pc.LocalAddr().String(), func() { _ = server.Shutdown() }
 }
 
 // TestRunCheck_CheckerError covers the checkErrCh error-wrapping path
